Resolve SonarCloud findings on the composition change (#1477)
- main.py: define _ENVIRONMENT_LOG constant instead of duplicating the literal.
- aces_range_ops: docstring on the inner registry resolver.
- Split aces_plan.py (>500 lines) -> aces_composition.py owns the content/feature/
  account value objects + extraction; aces_plan re-exports them. Fixes the walrus-
  in-argument-list finding by extracting the comprehensions to locals.
- Rename the account 'shell' field to 'login_shell', eliminating the bandit/ruff
  B604/S604 false positive (subprocess-shell heuristic) and its trailing nosec.
- Split runtime_target.py (>500 lines) -> composition_envelope.py owns the
  content/feature/account envelope diagnostics.
- Standalone provisioner-module loaders (drift guard + e2e) put the provisioner dir
  on sys.path so aces_plan's new sibling import resolves.

All local gates green: 102 provisioner + 182 platform ACES tests, ruff, mypy,
bandit, adr_guard.
