+++ shifter/engine/provisioner/aces_range_ops.py
@@ -34,6 +34,7 @@ def _registry_resolver() -> Callable[[AcesPlanNode], GCERangeImageProfile]:
     """Return an image resolver bound to the tenant registry + GCE policy."""
 
     def resolve(node: AcesPlanNode) -> GCERangeImageProfile:
+        """Resolve one node's image profile from the registry (authored source, else os_family)."""
         # Authored source keys the lookup; a source-less node falls back to its
         # os_family so the backend can supply a base OS image (ADR-032 base-OS policy).
         name = (node.image.name if node.image and node.image.name else node.os_family) or ""
