+++ shifter/engine/provisioner/main.py
@@ -16,6 +16,8 @@ from terraform_ops import run_range_terraform
 
 logger = logging.getLogger(__name__)
 
+_ENVIRONMENT_LOG = "Environment: %s"
+
 
 if __name__ == "__main__":
     from logging_config import configure_logging
@@ -82,7 +84,7 @@ if __name__ == "__main__":
         from aces_range_ops import run_aces_range_destroy, run_aces_range_provision
 
         logger.info("Starting ACES range %s for request_id=%s", args.operation, args.request_id)
-        logger.info("Environment: %s", os.environ.get("ENVIRONMENT", "unknown"))
+        logger.info(_ENVIRONMENT_LOG, os.environ.get("ENVIRONMENT", "unknown"))
         if args.operation == "provision":
             run_aces_range_provision(args.request_id)
         else:
@@ -91,7 +93,7 @@ if __name__ == "__main__":
 
     elif args.resource == "ngfw":
         logger.info("Starting NGFW %s for request_id=%s", args.operation, args.request_id)
-        logger.info("Environment: %s", os.environ.get("ENVIRONMENT", "unknown"))
+        logger.info(_ENVIRONMENT_LOG, os.environ.get("ENVIRONMENT", "unknown"))
 
         if args.operation in ("provision", "deprovision"):
             tf_op = "up" if args.operation == "provision" else "destroy"
@@ -109,7 +111,7 @@ if __name__ == "__main__":
         tf_op = "up" if args.operation == "provision" else "destroy"
 
         logger.info("Starting range %s for request_id=%s", args.operation, request_id)
-        logger.info("Environment: %s", os.environ.get("ENVIRONMENT", "unknown"))
+        logger.info(_ENVIRONMENT_LOG, os.environ.get("ENVIRONMENT", "unknown"))
 
         if args.operation in ("provision", "destroy"):
             run_range_terraform(tf_op, request_id)
